Move Countdown to separate component
